requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.115.12
google-auth==2.39.0
google-genai==1.12.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
loguru==0.7.3
MarkupSafe==3.0.2
motor==3.7.0
openai==1.76.0
packaging==25.0
passlib==1.7.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.12.1
python-dotenv==1.1.0
python-magic==0.4.27
python-http-client==3.3.7
python-multipart==0.0.20
pytz==2025.2
requests==2.32.3
rsa==4.9.1
sendgrid==6.12.0
sniffio==1.3.1
starkbank-ecdsa==2.2.0
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
Werkzeug==3.1.3